Added Altium schematic
I also started redoing the code. It is too old to figure out what I was
doing with it. Instead I will start again with the knowledge and
experience I have aquired since then.
I am also changing the chip from the PIC12F675 to the PIC24F08KL200.
Mostly because I want more pins to switch the grounds and toggle a
latching relay. Also partly because it is a 24F chip with power saving
features and an architecture that I am more familiar with.