Updated Pin Initialization
I used the functions to specifically initialize each of the pins. This
used to take up multiple lines and had different parts of the same pin
being updated all over the place. Now it is all done in a simple, single
location.

The pin names will be renamed to the schematic net names at a later date
to clarify the usage.