Incremental Change
I updated the timer initializations to use with the project, so far I
have the nessecary main loop regulation ready to go (but not implemented
yet.

I have expanded the project setup file to include the FOSC value as a
definition as well as defining the capture/compare modules (I know some
PICs have 8 of these, I don't know if this one does). I also added 4
UART modules, again same thing as the capture compares.

I removed the fan controller definition as this file is unique to the
project, so it does not need this kind of limit.